refactor: 돈 지불하고 구매 로또 개수 정하는 메서드 분리

+++ src/main/java/lotto/view/InputView.java
@@ -57,7 +57,7 @@ public class InputView {
         }
     }
 
-    public static List<String> askManualLottoNumbers(int count) {
+    public static List<String> askManualLottoNumbers(long count) {
         System.out.println(LINE_SEPARATOR + INPUT_MANUAL_LOTTO_NUMBERS);
         try {
             return Stream.generate(() -> SCANNER.nextLine())
